requirements.txt
streamlit>=1.32.0
gspread>=6.0.0
google-auth>=2.28.0
pandas>=2.0.0
numpy>=1.26.0
plotly>=5.18.0
python-dotenv>=1.0.0
